+++ Web-application/Dockerfile.dev
@@ -14,4 +14,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 EXPOSE 5000
 
-CMD ["flask", "--app", "flask_app", "run", "--host=0.0.0.0", "--port=5000", "--debug"]
+CMD ["flask", "--app", "flask_app.app", "run", "--host=0.0.0.0", "--port=5000", "--debug"]
