+++ Web-application/flask_app/templates/index.html
@@ -11,8 +11,8 @@
 
 <body>
     <main>
-        <a class="button" href="{{ url_for('login') }}">Login</a>
-        <a class="button" href="{{ url_for('register') }}">Register</a>
+        <a class="button" href="{{ url_for('auth.login') }}">Login</a>
+        <a class="button" href="{{ url_for('auth.register') }}">Register</a>
     </main>
 </body>
