+++ Web-application/flask_app/templates/login.html
@@ -32,7 +32,7 @@
                 </div>
             </form>
             <p class="status-msg"></p>
-            <a class="redirect-link" href="{{ url_for('recover') }}">Recover My Account</a>
+            <a class="redirect-link" href="{{ url_for('auth.recover') }}">Recover My Account</a>
         </div>
     </main>
     <script src="{{url_for('static', filename='js/index.js')}}"></script>
