+++ Web-application/flask_app/util/__init__.py
@@ -1,0 +1,5 @@
+from . import auth2
+from .recovery import verify_mnemonic, generate_mnemonic, hash_seed, convert_to_seed
+from .verify import verify
+from .qrGen import verify_totp, generate_qr
+from . import database
