+++ Web-application/flask_app/util/qrGen.py
